Remove STEP parsing features
